patch update. Refactored services and utilities
