+++ package.json
@@ -1,7 +1,7 @@
 {
   "private": true,
   "sideEffects": false,
-  "version": "0.1.0",
+  "version": "0.1.1",
   "scripts": {
     "build": "remix build",
     "dev": "remix dev"
