+++ src/pages/Home/loader.ts
@@ -1,6 +1,7 @@
 import { json } from '@remix-run/server-runtime'
-import { getGames } from 'utilities/api/service'
 import { LoaderArgs, SerializeFrom } from '@remix-run/node'
+
+import { getGames } from 'utilities/api/get-games-service'
 
 import dayjs from 'dayjs'
 
@@ -10,7 +11,6 @@ export const loader = async ({ request }: LoaderArgs) => {
   const url = new URL(request.url)
 
   const startDate = url.searchParams.get('startDate')
-
   const endDate = url.searchParams.get('endDate')
 
   if (startDate && endDate) {
